requirements.txt
langchain 
openai 
chromadb 
pinecone-client 
tiktoken
sentence-transformers
python-dotenv
PyPDF2
langchain-community
langchain-openai 
langchain-chroma